+++ src/main/java/de/samply/share/client/job/util/CqlResultFactory.java
@@ -2,10 +2,16 @@ package de.samply.share.client.job.util;
 
 import de.samply.share.client.model.db.tables.pojos.InquiryCriteria;
 import de.samply.share.client.model.db.tables.pojos.InquiryDetails;
+import de.samply.share.client.model.db.tables.pojos.InquiryResult;
 import de.samply.share.client.util.db.InquiryCriteriaUtil;
 import de.samply.share.client.util.db.InquiryResultUtil;
+import de.samply.share.model.common.result.Stratification;
 import de.samply.share.model.cql.CqlResult;
+import org.codehaus.jackson.map.ObjectMapper;
+import org.codehaus.jackson.type.TypeReference;
 
+import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
 public class CqlResultFactory {
@@ -18,19 +24,32 @@ public class CqlResultFactory {
 
     public CqlResult createCqlResult() {
         List<InquiryCriteria> inquiryCriteriaList = InquiryCriteriaUtil.getInquiryCriteriaForInquiryDetails(inquiryDetails);
-        int patientCount = 0;
-        int specimenCount = 0;
+        CqlResult queryResult = new CqlResult();
+
         for (InquiryCriteria inquiryCriteria : inquiryCriteriaList) {
+            InquiryResult inquiryResult = InquiryResultUtil.fetchLatestInquiryResultForInquiryCriteriaById(inquiryCriteria.getId());
             if (InquiryCriteriaEntityType.PATIENT.getName().equals(inquiryCriteria.getEntityType())) {
-                patientCount = InquiryResultUtil.fetchLatestInquiryResultForInquiryCriteriaById(inquiryCriteria.getId()).getSize();
+                queryResult.setNumberOfPatients(inquiryResult.getSize());
+                List<Stratification> stratifications = readStratifications(inquiryResult);
+                queryResult.setStratificationsOfPatients(stratifications);
             } else if (InquiryCriteriaEntityType.SPECIMEN.getName().equals(inquiryCriteria.getEntityType())) {
-                specimenCount = InquiryResultUtil.fetchLatestInquiryResultForInquiryCriteriaById(inquiryCriteria.getId()).getSize();
+                queryResult.setNumberOfSpecimens(inquiryResult.getSize());
+                List<Stratification> stratifications = readStratifications(inquiryResult);
+                queryResult.setStratificationsOfSpecimens(stratifications);
             }
         }
-        CqlResult queryResult = new CqlResult();
-        queryResult.setNumberOfPatients(patientCount);
-        queryResult.setNumberOfSpecimens(specimenCount);
+
         return queryResult;
+    }
+
+    private List<Stratification> readStratifications(InquiryResult inquiryResult) {
+        ObjectMapper mapper = new ObjectMapper();
+
+        try {
+            return mapper.readValue(inquiryResult.getStratifications(), new TypeReference<List<Stratification>>(){});
+        } catch (IOException e) {
+            return new ArrayList<>();
+        }
     }
 
 }