+++ src/main/resources/db/migration/V017__table_INQUIRY_RESULT_add_Stratifications.sql
@@ -1,0 +1,2 @@
+ALTER TABLE samply.inquiry_result
+ADD COLUMN stratifications text NOT NULL DEFAULT '[]';